requirements.txt
black==24.8.0
django==5.1
djangorestframework==3.15.2
flake8==7.1.1
isort==5.13.2
mypy==1.11.2